Rename libcamera-rs to libcamera
