+++ libcamera-sys/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "libcamera-sys"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
