+++ libcamera/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "libcamera-rs"
-version = "0.1.0"
+name = "libcamera"
+version = "0.2.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
 
@@ -8,7 +8,7 @@ license = "MIT OR Apache-2.0"
 bitflags = "2.0.0-rc.2"
 drm-fourcc = "2.2"
 libc = "0.2"
-libcamera-sys = { path = "../libcamera-sys" }
+libcamera-sys = { path = "../libcamera-sys", version = "0.2.0" }
 num_enum = "0.5"
 smallvec = "1.10"
 thiserror = "1.0"
