+++ libcamera/examples/video_capture.rs
@@ -1,6 +1,6 @@
-use std::time::Duration;
+use std::{fs::OpenOptions, io::Write, process::exit, time::Duration};
 
-use libcamera_rs::{
+use libcamera::{
     camera::CameraConfigurationStatus,
     camera_manager::CameraManager,
     framebuffer::AsFrameBuffer,
@@ -8,6 +8,7 @@ use libcamera_rs::{
     framebuffer_map::MemoryMappedFrameBuffer,
     pixel_format::PixelFormat,
     properties,
+    request::ReuseFlag,
     stream::StreamRole,
 };
 
@@ -15,7 +16,14 @@ use libcamera_rs::{
 const PIXEL_FORMAT_MJPEG: PixelFormat = PixelFormat::new(u32::from_le_bytes([b'M', b'J', b'P', b'G']), 0);
 
 fn main() {
-    let filename = std::env::args().nth(1).expect("Usage ./jpeg_capture <filename.jpg>");
+    let filename = match std::env::args().nth(1) {
+        Some(f) => f,
+        None => {
+            println!("Error: missing file output parameter");
+            println!("Usage: ./video_capture </path/to/output.mjpeg>");
+            exit(1);
+        }
+    };
 
     let mgr = CameraManager::new().unwrap();
 
@@ -31,9 +39,8 @@ fn main() {
     let mut cam = cam.acquire().expect("Unable to acquire camera");
 
     // This will generate default configuration for each specified role
-    let mut cfgs = cam.generate_configuration(&[StreamRole::ViewFinder]).unwrap();
+    let mut cfgs = cam.generate_configuration(&[StreamRole::VideoRecording]).unwrap();
 
-    // Use MJPEG format so we can write resulting frame directly into jpeg file
     cfgs.get_mut(0).unwrap().set_pixel_format(PIXEL_FORMAT_MJPEG);
 
     println!("Generated config: {:#?}", cfgs);
@@ -68,10 +75,11 @@ fn main() {
         .collect::<Vec<_>>();
 
     // Create capture requests and attach buffers
-    let mut reqs = buffers
+    let reqs = buffers
         .into_iter()
-        .map(|buf| {
-            let mut req = cam.create_request(None).unwrap();
+        .enumerate()
+        .map(|(i, buf)| {
+            let mut req = cam.create_request(Some(i as u64)).unwrap();
             req.add_buffer(&stream, buf).unwrap();
             req
         })
@@ -83,29 +91,47 @@ fn main() {
         tx.send(req).unwrap();
     });
 
+    // TODO: Set `Control::FrameDuration()` here. Blocked on https://github.com/lit-robotics/libcamera-rs/issues/2
     cam.start(None).unwrap();
 
-    // Multiple requests can be queued at a time, but for this example we just want a single frame.
-    cam.queue_request(reqs.pop().unwrap()).unwrap();
+    // Enqueue all requests to the camera
+    for req in reqs {
+        println!("Request queued for execution: {req:#?}");
+        cam.queue_request(req).unwrap();
+    }
 
-    println!("Waiting for camera request execution");
-    let req = rx.recv_timeout(Duration::from_secs(2)).expect("Camera request failed");
+    let mut file = OpenOptions::new()
+        .append(true)
+        .create(true)
+        .open(&filename)
+        .expect("Unable to create output file");
+    let mut count = 0;
+    while count < 60 {
+        println!("Waiting for camera request execution");
+        let mut req = rx.recv_timeout(Duration::from_secs(2)).expect("Camera request failed");
 
-    println!("Camera request {:?} completed!", req);
-    println!("Metadata: {:#?}", req.metadata());
+        println!("Camera request {:?} completed!", req);
+        println!("Metadata: {:#?}", req.metadata());
 
-    // Get framebuffer for our stream
-    let framebuffer: &MemoryMappedFrameBuffer<FrameBuffer> = req.buffer(&stream).unwrap();
-    println!("FrameBuffer metadata: {:#?}", framebuffer.metadata());
+        // Get framebuffer for our stream
+        let framebuffer: &MemoryMappedFrameBuffer<FrameBuffer> = req.buffer(&stream).unwrap();
+        println!("FrameBuffer metadata: {:#?}", framebuffer.metadata());
 
-    // MJPEG format has only one data plane containing encoded jpeg data with all the headers
-    let planes = framebuffer.data();
-    let jpeg_data = planes.get(0).unwrap();
-    // Actual JPEG-encoded data will be smalled than framebuffer size, its length can be obtained from metadata.
-    let jpeg_len = framebuffer.metadata().unwrap().planes().get(0).unwrap().bytes_used as usize;
+        // MJPEG format has only one data plane containing encoded jpeg data with all the headers
+        let planes = framebuffer.data();
+        let frame_data = planes.get(0).unwrap();
+        // Actual encoded data will be smalled than framebuffer size, its length can be obtained from metadata.
+        let bytes_used = framebuffer.metadata().unwrap().planes().get(0).unwrap().bytes_used as usize;
 
-    std::fs::write(&filename, &jpeg_data[..jpeg_len]).unwrap();
-    println!("Written {} bytes to {}", jpeg_len, &filename);
+        file.write(&frame_data[..bytes_used]).unwrap();
+        println!("Written {} bytes to {}", bytes_used, &filename);
+
+        // Recycle the request back to the camera for execution
+        req.reuse(ReuseFlag::REUSE_BUFFERS);
+        cam.queue_request(req).unwrap();
+
+        count += 1;
+    }
 
     // Everything is cleaned up automatically by Drop implementations
 }
