+++ libcamera/src/camera.rs
