+++ libcamera/src/camera_manager.rs
