+++ libcamera/src/control.rs
