+++ libcamera/src/control_value.rs
