+++ libcamera/src/framebuffer.rs
