+++ libcamera/src/framebuffer_allocator.rs
