+++ libcamera/src/framebuffer_map.rs
