+++ libcamera/src/generated.rs
