+++ libcamera/src/geometry.rs
