+++ libcamera/src/lib.rs
