+++ libcamera/src/stream.rs
