+++ libcamera/src/utils.rs
